+++ libcloud/compute/drivers/dimensiondata.py
@@ -538,6 +538,7 @@ class DimensionDataNodeDriver(NodeDriver):
             id=vlan_id,
             name=name,
             description=description,
+            network_domain=network_domain,
             location=network_domain.location,
             status=NodeState.RUNNING,
             private_ipv4_range_address=private_ipv4_base_address,
@@ -815,6 +816,35 @@ class DimensionDataNodeDriver(NodeDriver):
                 filter(lambda x: x.id == id, self.list_locations()))[0]
         return location
 
+    def ex_wait_for_state(self, state, func, poll_interval=2,
+                          timeout=60, *args, **kwargs):
+        """
+        Wait for the function which returns a instance
+        with field status to match
+
+        Keep polling func until one of the desired states is matched
+
+        :param state: Either the desired state (`str`) or a `list` of states
+        :type  state: ``str`` or ``list``
+
+        :param  func: The function to call, e.g. ex_get_vlan
+        :type   func: ``function``
+
+        :param  poll_interval: The number of seconds to wait between checks
+        :type   poll_interval: `int`
+
+        :param  timeout: The total number of seconds to wait to reach a state
+        :type   timeout: `int`
+
+        :param  args: The arguments for func
+        :type   args: Positional arguments
+
+        :param  kwargs: The arguments for func
+        :type   kwargs: Keyword arguments
+        """
+        return self.connection.wait_for_state(state, func, poll_interval,
+                                              timeout, *args, **kwargs)
+
     def _to_nat_rules(self, object, network_domain):
         rules = []
         for element in findall(object, 'natRule', TYPES_URN):
@@ -824,14 +854,12 @@ class DimensionDataNodeDriver(NodeDriver):
         return rules
 
     def _to_nat_rule(self, element, network_domain):
-        status = self._to_status(element.find(fixxpath('state', TYPES_URN)))
-
         return DimensionDataNatRule(
             id=element.get('id'),
             network_domain=network_domain,
             internal_ip=findtext(element, 'internalIp', TYPES_URN),
             external_ip=findtext(element, 'externalIp', TYPES_URN),
-            status=status)
+            status=findtext(element, 'state', TYPES_URN))
 
     def _to_firewall_rules(self, object, network_domain):
         rules = []
@@ -843,8 +871,6 @@ class DimensionDataNodeDriver(NodeDriver):
         return rules
 
     def _to_firewall_rule(self, element, locations, network_domain):
-        status = self._to_status(element.find(fixxpath('state', TYPES_URN)))
-
         location_id = element.get('datacenterId')
         location = list(filter(lambda x: x.id == location_id,
                                locations))[0]
@@ -862,7 +888,7 @@ class DimensionDataNodeDriver(NodeDriver):
             destination=self._to_firewall_address(
                 element.find(fixxpath('destination', TYPES_URN))),
             location=location,
-            status=status)
+            status=findtext(element, 'state', TYPES_URN))
 
     def _to_firewall_address(self, element):
         ip = element.find(fixxpath('ip', TYPES_URN))
@@ -884,8 +910,6 @@ class DimensionDataNodeDriver(NodeDriver):
         return blocks
 
     def _to_ip_block(self, element, locations):
-        status = self._to_status(element.find(fixxpath('state', TYPES_URN)))
-
         location_id = element.get('datacenterId')
         location = list(filter(lambda x: x.id == location_id,
                                locations))[0]
@@ -898,7 +922,7 @@ class DimensionDataNodeDriver(NodeDriver):
             base_ip=findtext(element, 'baseIp', TYPES_URN),
             size=findtext(element, 'size', TYPES_URN),
             location=location,
-            status=status)
+            status=findtext(element, 'state', TYPES_URN))
 
     def _to_networks(self, object):
         networks = []
@@ -939,8 +963,6 @@ class DimensionDataNodeDriver(NodeDriver):
         return network_domains
 
     def _to_network_domain(self, element, locations):
-        status = self._to_status(element.find(fixxpath('state', TYPES_URN)))
-
         location_id = element.get('datacenterId')
         location = list(filter(lambda x: x.id == location_id,
                                locations))[0]
@@ -955,7 +977,7 @@ class DimensionDataNodeDriver(NodeDriver):
             description=findtext(element, 'description', TYPES_URN),
             plan=plan_type,
             location=location,
-            status=status)
+            status=findtext(element, 'state', TYPES_URN))
 
     def _to_vlans(self, object):
         vlans = []
@@ -966,21 +988,24 @@ class DimensionDataNodeDriver(NodeDriver):
         return vlans
 
     def _to_vlan(self, element, locations):
-        status = self._to_status(element.find(fixxpath('state', TYPES_URN)))
-
         location_id = element.get('datacenterId')
         location = list(filter(lambda x: x.id == location_id,
                                locations))[0]
         ip_range = element.find(fixxpath('privateIpv4Range', TYPES_URN))
+        network_domain_el = element.find(
+            fixxpath('networkDomain', TYPES_URN))
+        network_domain = self.ex_get_network_domain(
+            network_domain_el.get('id'))
         return DimensionDataVlan(
             id=element.get('id'),
             name=findtext(element, 'name', TYPES_URN),
             description=findtext(element, 'description',
                                  TYPES_URN),
+            network_domain=network_domain,
             private_ipv4_range_address=ip_range.get('address'),
             private_ipv4_range_size=ip_range.get('prefixSize'),
             location=location,
-            status=status)
+            status=findtext(element, 'state', TYPES_URN))
 
     def _to_locations(self, object):
         locations = []